[SID:3579] refactor: 러너 정규화 60초 가드 경계대역 2파일 3-way 분할(코드 무변화)
PO 실측(2026-09-06) — develop 5212448ca 자체가 destructive 60초 가드 경계대역:
test_620beefc_channel_post_image.py(25건) 48~57s(판정선 51.7~64.0)·
test_3373_channel_connections.py(31건) 36~77s(판정선 56.0~68.0) — 러너가 조금만
느려져도 뒤집히는 자리(PR #3925/#3926에서는 각 174s/141s까지 관측). 3중 실측
(쿼리 카운트+스키마 생성+로컬 벽시계, #3922 전후 dabe141c2 vs 5212448ca 대조)
으로 코드 회귀 가설은 기각 — #3562 test_3414 분할 전례와 동일 처방(가드 무변·
주제별 파일 분할)으로 확定.

주제별 3-way 분할(테스트 본문 무변경, sed로 정확한 줄 범위 추출 — 손 재입력 0줄):
- `test_620beefc_channel_post_image_upload.py`(11건) — 업로드+변환+계보(AC1/AC2)·
  변환 불가 3종 422. 세팅 헬퍼·픽스처 전부 이 파일이 소유(원본 그대로) — 아래
  두 파일이 여기서 import.
- `test_620beefc_channel_post_image_lifecycle.py`(10건) — 봉인 판정 축 세분화
  (AC4 media 재승인)+비동기 발행 흐름 3갈래(AC5).
- `test_620beefc_channel_post_image_regression.py`(4건) — 텍스트 발행 회귀·
  어댑터 규격 노출(AC2)·create_only 서명(B4).
- `test_3373_channel_connections_auth.py`(13건) — crypto·OAuth state·agent
  403(AC6)·authorize→callback·disconnect(AC5). 세팅 헬퍼 소유.
- `test_3373_channel_connections_app_credentials.py`(14건) — cron 갱신(AC4)·
  평문 미노출 grep·조직별 앱 자격·3단 우선순위·effective_source.
- `test_3373_channel_connections_pkce.py`(4건) — PKCE 플래그 on/off.

autouse 픽스처(`_dispose_global_engine_after_test`·`_configure_secrets`·
620beefc 쪽 `_local_channel_media_storage`)는 pytest 관례상 import로 안
전파돼 파일마다 재선언(관례 그대로, 새 프리미티브 0, #3562 전례와 동일).

의존 파일 갱신(신규 헬퍼 0, import 경로만 새 파일명으로) — 실제 코드 import를
쓰던 test_3320/test_3550/test_3554/test_3559/test_3567_facebook_page_final.py
(`from tests.X import`·`import tests.X as base_test_module` 두 형태 전부)·
test_3523/test_3547(test_3373 재사용). 프로스만 인용하던 test_3399/
test_5b27b32f/test_f8f7cb0f도 새 파일명으로 갱신(코드 import는 원래 0건).
`test_shard_destructive_tests.py`의 옛 CI run(33773872963) 냉동 표본 데이터는
역사적 스냅샷이라 그대로 둠(리네임 대상 아님).

weights.json: 원본 2항목 제거, 6개 신규 항목 등록(원본 등재값을 테스트 개수/
로컬 raw pytest 시간 비중으로 배분한 잠정값 — source에 명시, 실 CI 샤드 로그
재확認 필요).

검증: assert 개수 대조(620beefc 99=99·3373 100=100, 손실/중복 0)·전체
destructive_schema 컬렉션 에러 0(2108건)·직접 관련 파일 회귀 220건 전체 통과·
shard_destructive_tests 관련 lint 3종 통과·weights 등재 lint 통과. base develop.
